Extract Lua events to its own source file
