+++ src/lua/luascript.h
@@ -4,20 +4,11 @@
 
 #include "types.h"
 
-typedef enum  {
-  LUA_EVENT_OPEN,
-  LUA_EVENT_CLOSE,
-  LUA_EVENT_REPORT,
-  LUA_EVENT_ITERATION,
-  LUA_EVENT_MAX
-} LuaEvent;
-
 int luascript_open(Project *pr);
 void luascript_close(Project *pr);
 
 int luascript_addScriptLine(Project *pr, char *line);
 int luascript_parseScript(Project *pr);
-int luascript_onEvent(Project *pr, LuaEvent event);
 void luascript_setChanged(Project *pr);
 
 #endif // LUA_SCRIPTING
